- Context.java Added closedSetIceberg method (it surely can be extended to any kind of ClosureSystem, but I don't know how for now).
- ConceptLattice.java
diagramIceberg() and recursiveDiagramIceberg() are derived from their Lattice counterparts.
They are called from Context's closedSetIceberg.
These methods are to be prefered over ConceptLattice's iceberg() as they do not need the full concept lattice to be generated to begin with.
Thus, iceberg() method is now deprecated.

- ImplicationalSystem.java
Modified makeCompact() to make it faster. Performances are increased by a factor I would compute if I had the time to wait for the original method to finish...