Implementation of association rules generation
-AssociationRule.java
A few modifications in parameter order in the constructor and toString() methods in order to be consistent with some resource material.

-Lattice.java
getAssociationBasis(context, support, confidence) returns a set of approximative association rules, based on a couple of  support and confidence thresholds.

-ImplicationalSystem.java
makeCompactAssociation() merge rules with same premisse, same support and confidence values into a single one.